+++ QuickFIXn/Session.cs
@@ -1646,9 +1646,12 @@ public class Session : IDisposable
         if (PersistMessages)
         {
             SeqNumType msgSeqNum = message.Header.GetULong(Fields.Tags.MsgSeqNum);
-            _state.Set(msgSeqNum, messageString);
+            _state.SetAndIncrNextSenderMsgSeqNum(msgSeqNum, messageString);
         }
-        _state.IncrNextSenderMsgSeqNum();
+        else
+        {
+            _state.IncrNextSenderMsgSeqNum();
+        }
     }
 
     protected bool IsGoodTime(Message msg)