+++ QuickFIXn/SessionState.cs
@@ -373,6 +373,11 @@ public class SessionState : IDisposable
         lock (_sync) { MessageStore.IncrNextTargetMsgSeqNum(); }
     }
 
+    public bool SetAndIncrNextSenderMsgSeqNum(SeqNumType msgSeqNum, string msg)
+    {
+        lock (_sync) { return MessageStore.SetAndIncrNextSenderMsgSeqNum(msgSeqNum, msg); }
+    }
+
     public DateTime? CreationTime
     {
         get